Extract common AutoinstDevicesPlanner to AutoinstPlanner
